pyproject.toml
[tool.poetry]
name = "fidowinter"
version = "0.1.0"
description = "Notify mailing list of first day of winter"
authors = ["Stig Ofstad <stigofstad@gmail.com>"]
license = "GNU Affero General Public License v3.0"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.24.0"
google-api-python-client = "^1.12.3"
google-auth-oauthlib = "^0.4.1"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
